requirements.txt
pyglet==1.2.4
cocos2D